% Generated by roxygen2 (4.0.0): do not edit by hand
\name{exact_log_rank}
\alias{exact_log_rank}
\title{function for exact log-rank hypothesis tests}
\usage{
exact_log_rank(mod)
}
\arguments{
  \item{mod}{the summary of a survfit object from the
  survival package no other params just test}
}
\description{
function for exact log-rank hypothesis tests
}
\examples{
mod <- summary(survfit(Surv(time,ind)~ grp))
exact_log_rank(mod)
}
\keyword{log-rank}
\keyword{survival}
\keyword{test}

