% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ggHeat}
\alias{ggHeat}
\title{ggplot heatmap, ordered according to native 'heatmap dendrogram}
\usage{
ggHeat(bio4)
}
\arguments{
  \item{bio4}{matrix or dataframe whose heatmap you would
  like to plot}
}
\description{
ggplot heatmap, ordered according to native 'heatmap dendrogram
}
\examples{
g <- ggheat(bio4)
print(g)
}
\keyword{correlation}
\keyword{covariance}
\keyword{heatmap}

