% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ggSurv}
\alias{ggSurv}
\title{ggplot heatmap, ordered according to native 'heatmap dendrogram}
\usage{
ggSurv(EventTime, EventStatus, nam, dat)
}
\arguments{
  \item{surv}{EventTime survival times EventStatus
  censoring indicator (1 implies event observed, 0
  otherwise) nam string literal of dat indicating variable
  by which KM curves should be stratified dat dataframe
  with above 3 mentioned covariates}
}
\description{
ggplot heatmap, ordered according to native 'heatmap dendrogram
}
\examples{
g <- ggSurv(EventTime,EventStatus,'plink_rs2716212_G',dat)
ggsave(g,file='~/test_gg.jpg',height=6,width=8)
}
\keyword{curve}
\keyword{kaplan-meier}
\keyword{km}
\keyword{survival}

